PredefinedNoteService Refine, NoteServiceUnitTests

+++ src/main/java/com/project/mini/service/PredefinedNotesService.java
@@ -41,7 +41,7 @@ public class PredefinedNotesService {
                 .map(predfniedNotes -> predfniedNotes.toDTO())
                 .collect(Collectors.toList());
     }
-    public String getPreefinedNote(double temperaure) {
+    public String getPreefinedNote(Float temperaure) {
         List<PredfinedNotesDTO> allPredefinedNotes = this.getAllPredfniedNotes();
         for (PredfinedNotesDTO note : allPredefinedNotes) {
             if(temperaure >= note.getMinimumTemperture() && note.getMaximumTemperture()> temperaure)
